+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.admin.v1/src/main/java/org/wso2/carbon/apimgt/rest/api/admin/v1/utils/RestApiAdminUtils.java
@@ -17,13 +17,31 @@
 
 package org.wso2.carbon.apimgt.rest.api.admin.v1.utils;
 
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.wso2.carbon.apimgt.api.APIManagementException;
 import org.wso2.carbon.apimgt.api.model.BlockConditionsDTO;
 import org.wso2.carbon.apimgt.api.model.policy.Policy;
 import org.wso2.carbon.apimgt.impl.utils.APIUtil;
 import org.wso2.carbon.utils.multitenancy.MultitenantUtils;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
+
 public class RestApiAdminUtils {
+
+    //using a set for file extensions white list since it will be faster to search
+    private static final Set<String> EXTENSION_WHITELIST = new HashSet<String>(Arrays.asList(
+            "css", "jpg", "png", "gif", "svg", "ttf", "html", "js", "json", "ico"));
 
     /**
      * Checks whether given policy is allowed to access to user
@@ -54,5 +72,90 @@ public class RestApiAdminUtils {
         String userTenantDomain = MultitenantUtils.getTenantDomain(user);
         return !StringUtils.isBlank(blockCondition.getTenantDomain()) && blockCondition.getTenantDomain()
                 .equals(userTenantDomain);
+    }
+
+    /**
+     * Extract the content of the provided tenant theme archive
+     *
+     * @param themeFile    content relevant to the tenant theme
+     * @param tenantDomain tenant to which the theme is imported
+     * @throws APIManagementException if an error occurs while importing tenant theme
+     */
+    public static void deployTenantTheme(InputStream themeFile, String tenantDomain) throws APIManagementException {
+
+        ZipInputStream zipInputStream = null;
+        byte[] buffer = new byte[1024];
+
+        String outputFolder = "repository" + File.separator + "deployment" + File.separator + "server"
+                + File.separator + "jaggeryapps" + File.separator + "devportal" + File.separator + "site"
+                + File.separator + "public" + File.separator + "tenant_themes" + File.separator + tenantDomain;
+
+        try {
+            //create output directory if it does not exist
+            File folder = new File(outputFolder);
+            if (!folder.exists()) {
+                if (!folder.mkdirs()) {
+                    APIUtil.handleException("Unable to create tenant theme directory at " + outputFolder);
+                }
+            } else {
+                //remove existing files inside the directory
+                FileUtils.cleanDirectory(folder);
+            }
+
+            //get the zip file content
+            zipInputStream = new ZipInputStream(themeFile);
+            //get the zipped file list entry
+            ZipEntry zipEntry = zipInputStream.getNextEntry();
+
+            while (zipEntry != null) {
+
+                String fileName = zipEntry.getName();
+                APIUtil.validateFileName(fileName);
+                File newFile = new File(outputFolder + File.separator + fileName);
+                String canonicalizedNewFilePath = newFile.getCanonicalPath();
+                String canonicalizedDestinationPath = new File(outputFolder).getCanonicalPath();
+                if (!canonicalizedNewFilePath.startsWith(canonicalizedDestinationPath)) {
+                    APIUtil.handleException(
+                            "Attempt to upload invalid zip archive with file at " + fileName + ". File path is " +
+                                    "outside target directory");
+                }
+
+                if (zipEntry.isDirectory()) {
+                    if (!newFile.exists()) {
+                        boolean status = newFile.mkdir();
+                        if (!status) {
+                            APIUtil.handleException("Error while creating " + newFile.getName() + " directory");
+                        }
+                    }
+                } else {
+                    String ext = FilenameUtils.getExtension(zipEntry.getName());
+                    if (EXTENSION_WHITELIST.contains(ext)) {
+                        //create all non exists folders
+                        //else you will hit FileNotFoundException for compressed folder
+                        new File(newFile.getParent()).mkdirs();
+                        FileOutputStream fileOutputStream = new FileOutputStream(newFile);
+
+                        int len;
+                        while ((len = zipInputStream.read(buffer)) > 0) {
+                            fileOutputStream.write(buffer, 0, len);
+                        }
+
+                        fileOutputStream.close();
+                    } else {
+                        APIUtil.handleException(
+                                "Unsupported file is uploaded with tenant theme by " + tenantDomain + " : file name : "
+                                        + zipEntry.getName());
+                    }
+                }
+                zipEntry = zipInputStream.getNextEntry();
+            }
+            zipInputStream.closeEntry();
+            zipInputStream.close();
+        } catch (IOException ex) {
+            APIUtil.handleException("Failed to deploy tenant theme for tenant " + tenantDomain, ex);
+        } finally {
+            IOUtils.closeQuietly(zipInputStream);
+            IOUtils.closeQuietly(themeFile);
+        }
     }
 }
