+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.util/src/main/java/org/wso2/carbon/apimgt/rest/api/util/RestApiConstants.java
@@ -146,7 +146,7 @@ public final class RestApiConstants {
             REST_API_PUBLISHER_CONTEXT + REST_API_PUBLISHER_VERSION_0;
     public static final String REST_API_PUBLISHER_CONTEXT_FULL_1 =
             REST_API_PUBLISHER_CONTEXT + REST_API_PUBLISHER_VERSION_1;
-    public static final String REST_API_ADMIN_CONTEXT = "/api/am/admin";
+    public static final String REST_API_ADMIN_CONTEXT = "/api/am/admin/";
     public static final String REST_API_ADMIN_VERSION_0 = "v0.16";
     public static final String REST_API_ADMIN_VERSION_1 = "v1";
     public static final String REST_API_ADMIN_CONTEXT_FULL_0 = REST_API_ADMIN_CONTEXT + REST_API_ADMIN_VERSION_0;
